Adding integration to gitignore
The script will check if the current directory is version controlled and
will append the generated .asset-config.json to .gitignore. This is
necessary to avoid placing any credential keys on github and prevent
their subsquential linking.
The user can still overide this, but the intention is for that not to
append.
Next step is to encrypt the configuration file, along side with avoiding
the file to be staged in git.